# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basalt API",
    "version": "1.1.0",
    "description": "Pro-tier hosted backend for Basalt. Runs on Cloudflare Workers; Open-tier surfaces (CLI, plugin, MCP, desktop) do not call this API.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/plsft/basalt/blob/main/LICENSE"
    },
    "contact": {
      "name": "Basalt",
      "url": "https://basalt.dev"
    }
  },
  "servers": [
    {
      "url": "https://api.basalt.dev",
      "description": "Production"
    },
    {
      "url": "https://api-staging.basalt.dev",
      "description": "Staging"
    }
  ],
  "components": {
    "schemas": {
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "tier": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "founder"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "tier"
        ]
      },
...
      "Vault": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sync_enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "created_at",
          "updated_at"
        ]
      },
      "VaultList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Vault"
        }
      },
      "VaultCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "upload",
              "sync"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Brief": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "integer",
            "const": 1
          },
          "section": {
            "type": "string"
          },
          "track_record": {
            "type": "object"
          },
          "findings": {
            "type": "object"
          }
        },
        "required": [
          "schema",
          "section",
          "findings"
        ]
      },
      "BriefList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Brief"
            }
          },
          "cursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
      "FindingList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "cursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items"
        ]
      },
...
      "ByokPutRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "google"
            ]
          },
          "api_key": {
            "type": "string",
            "minLength": 8,
            "maxLength": 4096
          }
        },
        "required": [
          "provider",
          "api_key"
        ]
      },
      "ByokStatusList": {
        "type": "object",
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "provider": {
                  "type": "string"
                },
                "has_key": {
                  "type": "boolean"
                }
              },
              "required": [
                "provider",
                "has_key"
              ]
            }
          }
        },
        "required": [
          "providers"
        ]
      },
...
      "SearchRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "vault_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "top": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 10
          }
        },
        "required": [
          "query"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          },
          "elapsed_ms": {
            "type": "integer"
          },
          "embedding_model": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "vault_id": {
                  "type": "string"
                },
                "rel_path": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "updated": {
                  "type": "string",
                  "nullable": true
                },
                "score": {
                  "type": "number"
                }
              },
              "required": [
                "vault_id",
                "rel_path",
                "title",
                "score"
              ]
            }
          }
        },
        "required": [
          "query",
          "elapsed_ms",
          "embedding_model",
          "hits"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "basalt_session",
        "description": "HttpOnly session cookie set by /v1/auth/oauth/callback."
      }
    }
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "summary": "Current user",
        "description": "Returns the authenticated user's id, email, and tier.",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "summary": "Soft-delete the current user",
        "description": "Sets deleted_at on the user row; data is purged after a 30-day grace window.",
        "tags": [
          "Identity"
        ],
        "responses": {
          "200": {
            "description": "Deletion scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/v1/vaults": {
      "get": {
        "summary": "List the user's vaults",
        "description": "Excludes soft-deleted vaults.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "200": {
            "description": "Vault list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultList"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "summary": "Register a new vault",
        "description": "Creates a vault row and enqueues an indexing job.",
        "tags": [
          "Vaults"
        ],
        "responses": {
          "201": {
            "description": "Vault created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vault"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaultCreateRequest"
              }
            }
          }
        }
      }
    },
...
    "/v1/search": {
      "post": {
        "summary": "Cross-vault semantic search",
        "description": "Embeds the query via Workers AI and queries Vectorize. Filters to user-owned vaults; vault_ids[] narrows further. Returns hits sorted by similarity.",
        "tags": [
          "Default"
        ],
        "responses": {
          "200": {
            "description": "Search hits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "x-rate-limit": {
          "scope": "search",
          "max": 60,
          "window": 60
        }
      }
    },
...
    "/v1/briefs": {
      "get": {
        "summary": "List briefs for a vault",
        "description": "Cursor-paginated.",
        "tags": [
          "Briefs"
        ],
        "responses": {
          "200": {
            "description": "Brief list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefList"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/findings": {
      "get": {
        "summary": "Cross-brief finding timeline",
        "description": "Filterable by verb, vault_id, status. Cursor-paginated by finding id.",
        "tags": [
          "Findings"
        ],
        "responses": {
          "200": {
            "description": "Findings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingList"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/v1/byok": {
      "get": {
        "summary": "List BYOK provider key status",
        "description": "Returns has_key per provider (openai, anthropic, google).",
        "tags": [
          "Default"
        ],
        "responses": {
          "200": {
            "description": "BYOK status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ByokStatusList"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "summary": "Store or rotate a BYOK provider key",
        "description": "AES-GCM encrypted at rest in BYOK_KEYS KV. Server-side master key only.",
        "tags": [
          "Default"
        ],
        "responses": {
          "200": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ByokPutRequest"
              }
            }
          }
        }
      }
    },
...
  }
}